2010-05-27	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libfs/src/imfs/imfs.h, libfs/src/imfs/imfs_eval.c:
	Use size_t instead of int for pathnamelen.

+++ cpukit/libfs/src/imfs/imfs.h
@@ -324,7 +324,7 @@ extern int IMFS_evaluate_link(
 
 extern int IMFS_eval_path(
   const char                        *pathname,     /* IN     */
-  int                               pathnamelen,   /* IN     */
+  size_t                            pathnamelen,   /* IN     */
   int                               flags,         /* IN     */
   rtems_filesystem_location_info_t  *pathloc       /* IN/OUT */
 );
